% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_summary.R
\name{get_db_summary}
\alias{get_db_summary}
\title{Get database summary}
\usage{
get_db_summary()
}
\value{
Data frame: Columns: "data type", "value".
}
\description{
Get a summary of the data stored.
}
\examples{

get_db_summary()

}
