% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_id_from_genome_seq.R
\name{get_genome_id_from_genome_seq}
\alias{get_genome_id_from_genome_seq}
\title{Get genome from genome sequence}
\usage{
get_genome_id_from_genome_seq(genome_seq)
}
\arguments{
\item{genome_seq}{String of letters or a list of strings}
}
\value{
Data frame. Columns: "genome_id" "genome_seq"
}
\description{
Get the genome of a digital organism from the linear string of
letters representing the instruction codes that make up its genome.
}
\examples{
# Get sequence for genome_1
sequence <- get_genome_seq_from_genome_id(genome_id = 1)$genome_seq[1]

#
get_genome_id_from_genome_seq(genome_seq = sequence)

}
