% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_seq_from_genome_id.R
\name{get_genome_seq_from_genome_id}
\alias{get_genome_seq_from_genome_id}
\title{Get genome sequence from genome}
\usage{
get_genome_seq_from_genome_id(genome_id)
}
\arguments{
\item{genome_id}{Integer or a list of integer values.}
}
\value{
Data frame. Columns: "genome_id" "genome_seq"
}
\description{
Get the linear string of letters representing the instruction
codes that make up the genome of a digital organism from the id of the genome
of a digital organism.
}
\examples{

# Single genome
get_genome_seq_from_genome_id(1)

# More than one genome
get_genome_seq_from_genome_id(genome_id = c(1, 2, 3, 4))

}
