% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phenotype_id_from_logic_operation.R
\name{get_phenotype_id_from_logic_operation}
\alias{get_phenotype_id_from_logic_operation}
\title{Get phenotype from logic operations}
\usage{
get_phenotype_id_from_logic_operation(
  logic_operation,
  phenotype_binary = FALSE
)
}
\arguments{
\item{logic_operation}{List of logical functions from the following set:
"equals", "exclusive-or", "not-or", "and-not", "or", "orn-not", "and",
"not-and", "not".}

\item{phenotype_binary}{Logical value (TRUE/FALSE) to show/hide the phenotype in binary notation (FALSE by default).}
}
\value{
Data frame. Columns: "phenotype_id", "phenotype_binary" (optional).
}
\description{
Get the phenotype encoded by the genome of a digital organism that is specified by a unique combination of logic operations.
}
\examples{

# Single logic operation
get_phenotype_id_from_logic_operation(logic_operation = "not-or")

# More than one logic operation
ops <- c("equals", "exclusive or", "not-or", "and-not", "or", "orn-not")
get_phenotype_id_from_logic_operation(
  logic_operation = ops,
  phenotype_binary = TRUE
)

}
