% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avidaR.R
\name{instruction_set}
\alias{instruction_set}
\title{Get the genetic language of Avida}
\usage{
instruction_set(inst_set = "heads")
}
\arguments{
\item{inst_set}{Name of the instruction set. It must be one of the
following: "heads" (default), "heads-sex", or \code{transsmt}. The names
correspond to the instruction set configuration files (e.g.,
\code{instset-heads.cfg} for "heads").}
}
\value{
Data frame. Columns: "instruction", "letter", "color"
}
\description{
List of the instruction codes comprising the genetic language of
digital organisms in Avida.
}
