% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avidaR.R
\name{logic_operation}
\alias{logic_operation}
\title{Get the list of logic operations that a digital organism can compute}
\usage{
logic_operation()
}
\value{
Vector of character.
}
\description{
List of the logic operations that a digital organism can
execute.
}
