% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_id_of_wild-type_organisms.R
\name{get_genome_id_of_wild_type_organisms}
\alias{get_genome_id_of_wild_type_organisms}
\title{Get genomes of wild-type organisms}
\usage{
get_genome_id_of_wild_type_organisms(triplestore)
}
\arguments{
\item{triplestore}{Object of class triplestore_access which manages database
access.}
}
\value{
Data frame: "genome_id_wild_type".
}
\description{
Get the genome of the digital organisms that were used as
wild-type organisms to get their single-point mutants by calling the function
get_mutant_at_pos.
}
\examples{

# Create triplestore object
avidaDB <- triplestore_access$new()

# Set access options
avidaDB$set_access_options(
  url = "https://graphdb.fortunalab.org",
  user = "public_avida",
  password = "public_avida",
  repository = "avidaDB_test"
)

get_genome_id_of_wild_type_organisms(triplestore = avidaDB)

}
