% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tandem_id_from_phenotype_id.R
\name{get_tandem_id_from_phenotype_id}
\alias{get_tandem_id_from_phenotype_id}
\title{Get tandem repeat from phenotype}
\usage{
get_tandem_id_from_phenotype_id(
  phenotype_id,
  seed_id = FALSE,
  tandem_seq = FALSE,
  tandem_pos = FALSE
)
}
\arguments{
\item{phenotype_id}{Integer or list of integer values.}

\item{seed_id}{Integer (from 1 to 1000), a vector of integer values, or a
logical value. This integer is used for starting the pseudo-random number
generator that represents the environment experiencing a digital organism.
If a logical value is used, TRUE returns data found in all environments and
FALSE (by default) returns only distinct data regardless of the seed.}

\item{tandem_seq}{Logical value (TRUE/FALSE) to show/hide this column
(FALSE by default).}

\item{tandem_pos}{Logical value (TRUE/FALSE) to show/hide this column
(FALSE by default).}
}
\value{
Data frame. Column: "seed_id" (optional), "phenotype_id",
"tandem_id", "tandem_seq" (optional), "tandem_pos (optional)."
}
\description{
Get the tandem repeat contained in the transcriptome of a
digital organism that encodes a specific phenotype for a list of seeds used
for starting the pseudo-random number generator (i.e., a set of environments).
}
\examples{

# Single phenotype
get_tandem_id_from_phenotype_id(phenotype_id = 8, tandem_seq = TRUE)

# More than one phenotype at seed_1
get_tandem_id_from_phenotype_id(phenotype_id = c(2, 4, 8), seed_id = 1)

# At seed_1 and seed_2
get_tandem_id_from_phenotype_id(
  phenotype_id = 1,
  seed_id = c(1, 2),
  tandem_pos = TRUE
)

}
