% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avidaR.R
\name{instruction_set}
\alias{instruction_set}
\title{Get the genetic language of Avida}
\usage{
instruction_set()
}
\value{
Data frame. Columns: "instruction", "letter", "color"
}
\description{
List of the instruction codes comprising the genetic language of
digital organisms in Avida.
}
